requirements.txt
# Web Scraping & Data Collection
requests
playwright
scrapy
lxml
trafilatura
beautifulsoup4

# Database
psycopg2-binary

# Feature Engineering & Exploratory Data Analysis 
scikit-learn
numpy
scipy
pandas
matplotlib
seaborn
notebook
jupyterlab
IPython